// src/app/models/usuario/byuser-model.ts
// Generated by https://quicktype.io

export interface ByUserModel {
    id:              string;
    nombreCompleto:  string;
    nombre:          string;
    apellido:        string;
    puesto:          string;
    telefono:        string;
    correo:          string;
    cuentaMaestra:   boolean;
    ciudad:          null;
    numAcredor:      string;
    descripcion:     null;
    mensajes:        any[];
    tipo:            Tipo;
    permisos:        Permiso[];
    razonesSociales: RazonesSociale[];
}

export interface Permiso {
    id:            number;
    nombre:        string;
    tipo:          Tipo;
    valor:         string;
    modulo:        Modulo;
    cuentaMaestra: number;
}

export enum Modulo {
    Empty = "",
    Facturación = "Facturación",
}

export enum Tipo {
    Mercaderia = "Mercaderia",
}

export interface RazonesSociale {
    id:             string;
    numProveedor:   string;
    razonSocial:    string;
    nombre:         string;
    email:          string;
    vendedor:       string;
    estado:         string;
    poblacion:      string;
    cp:             string;
    direccion:      string;
    centroDeCostos: string;
    cuentaMayor:    string;
    sociedades:     Sociedade[];
}

export interface Sociedade {
    id:       string;
    nombre:   string;
    sociedad: string;
    activo:   boolean;
}